back to original openmc with trigger

+++ src/utils/particle_restart.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 
 import struct
 
@@ -50,9 +50,9 @@ class Particle(object):
 
     def _get_long(self, n=1, path=None):
         if self._hdf5:
-            return [int(v) for v in self._f[path].value]
+            return [long(v) for v in self._f[path].value]
         else:
-            return [int(v) for v in self._get_data(n, 'q', 8)]
+            return [long(v) for v in self._get_data(n, 'q', 8)]
 
     def _get_float(self, n=1, path=None):
         if self._hdf5:
